Move detail namespace stuff to separate file
So that the important class is right at the top.

---
